consolidated about class code
